+++ packages/wordpress-kit/CHANGELOG.md
@@ -1,10 +1,17 @@
 # @pantheon-systems/wordpress-kit
 
+## 2.1.0
+
+### Minor Changes
+
+- 0db7b52: Created a plugin for tailwind, that supports the block editor classes
+
 ## 2.0.0
 
 ### Major Changes
 
-- 5dfbb69: [wordpress-kit] We move to use graphql-request instead of apollo client, updated the test for that, the main example and the exports
+- 5dfbb69: [wordpress-kit] We move to use graphql-request instead of apollo
+  client, updated the test for that, the main example and the exports
 
   ### Breaking